Throw error when 'alt' is given no alternates
The previous behaviour was to fail with an error during parsing if there
are no alternates.  I can't see a situation in which users would
actually want that.

I also considered making 'alt' with no alternates just succeed with
nothing consumed, but that would result in hard-to-debug infinite loops
if someone passes an empty 'alt' into 'times' for example, and even so
it isn't really clear that that would be semantically correct at all.