Make p.seq take array, and add env chaining to it
I have this new idea I want to try.  It requires seq parsers to be
passed in an array, so another parameter can be passed.  It's for
user-determined environment passing between parsers in the sequence!
Cool in case you want to use an immutable environment that gets passed
back down as part of the parse result.  This way, you can also pass it
onward through a sequence of parsers.